ceph_diagnostics_collect: add scrip version information
Its is not actually a version but just md5sum of the scrip file. But
collecting this info is useful to understand if a user was running the
latest script version when collecting data. For this one may compare
the contents of script_info-version with hers output of
`./ceph_diagnostics_collect.sh --version`.

Signed-off-by: Mykola Golub <mykola.golub@clyso.com>