ceph_diagnostics_collect: create COMMANDS symlinks
A symlink has the same name as the corresponding collect
command. E.g. if `ceph df` command was run to collect output in
cluster_health-df file, there will also be created a symlink
"COMMANDS/ceph df" for this file.

This is supposed to be useful for analysing tools like cds.

Signed-off-by: Mykola Golub <mykola.golub@clyso.com>